[MARKENG-1080] Developed Modal functionality for images to open on user click if it has an <a> tag in markdown
