+++ src/templates/doc.jsx
@@ -10,10 +10,10 @@ import SEO from '../components/seo';
 import './doc.scss';
 import 'prismjs/themes/prism-tomorrow.css';
 import pose from '../assets/pose-learning-center.svg';
-
+import { useModal } from '../components/Modal/useModal.js';
 const { v4: uuidv4 } = require('uuid');
 
-class GenerateDoc extends React.Component {
+class CreateDoc extends React.Component {
   constructor(props) {
     super(props);
     const { data } = this.props;
@@ -23,72 +23,17 @@ class GenerateDoc extends React.Component {
   }
 
   componentDidMount() {
-
     let { data } = this.state;
     const { html } = data.markdownRemark;
     let parser = new DOMParser();
     let htmlDocument = parser.parseFromString(html, 'text/html');
-    let links = htmlDocument.querySelectorAll('img');
-    const create_id = `img-${Math.random().toString(36).slice(2)}`
-    for (let i = 0; i < links.length; i++) {
-      if( links[i].src === links[i].parentNode.href){
-        links[i].parentNode.href = "javascript:void(0)";
-        links[i].outerHTML = (
-          `<>
-          <a data-target=#${create_id} class="modal-image" data-toggle="modal">
-            <img src=${links[i].src}  />
-          </a>
-          <span id=${create_id} class="modal fade modal-backdrop" tabIndex="-1" role="dialog" aria-labelledby={#${create_id}} aria-hidden="true">
-            <span class="container modal-dialog modal-dialog-centered" role="document">
-              <span class="modal-row row">
-                <span class="col-8">
-                  <span class="modal-close-button">
-                    <button
-                      type="button"
-                      class="close"
-                      data-dismiss="modal"
-                      aria-label="Close"
-                    >
-                      <span
-                        class='close-icon'
-                        aria-hidden="true"
-                      >
-                        <svg
-                          xmlns="http://www.w3.org/2000/svg"
-                          width="16"
-                          height="16"
-                          viewBox="0 0 16 16"
-                        >
-                          <title>e-remove</title>
-                          <g
-                            strokeWidth="1"
-                            fill="#212121"
-                            stroke="#212121"
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                          >
-                            <line x1="13.5" y1="2.5" x2="2.5" y2="13.5" />
-                            <line x1="2.5" y1="2.5" x2="13.5" y2="13.5" />
-                          </g>
-                        </svg>
-                      </span>
-                    </button>
-                  </span>
-                  <img src=${links[i].src}  />
-                </span>
-              </span>
-            </span>
-          </span>
-        </>`
-        )
-      } 
-    }
-    
+    /* Import JS scripts to render a component on the doc page /*
+    /* Enables Images to Use Modal (returns htmlDocument) */
+    useModal(htmlDocument);
 
     this.setState({
       htmlDocument: htmlDocument.body.innerHTML
     });
-
   }
 
   render() {
@@ -99,9 +44,7 @@ class GenerateDoc extends React.Component {
   }
 }
 
-
-
-
+// Layout for Doc Page
 const DocPage = ({ data }) => {
   const post = data.markdownRemark;
   let contextualLinks;
@@ -121,7 +64,7 @@ const DocPage = ({ data }) => {
             <div className="row row-eq-height">
               <main className="col-sm-12 col-md-12 col-lg-9 offset-lg-0 col-xl-7 doc-page ml-xl-5">
                 <h1>{post.frontmatter.title}</h1>
-                <GenerateDoc data={data}/>
+                <CreateDoc data={data}/>
               </main>
               <aside className="col-sm-12 col-md-12 col-lg-3 offset-lg-0 col-xl-3 offset-xl-1 right-column">
                 <hr className="d-block d-lg-none" />
